Removing two casts "(long)" in fmpz/pow_ui.c

+++ fmpz/pow_ui.c
@@ -43,8 +43,8 @@ fmpz_pow_ui(fmpz_t f, const fmpz_t g, ulong exp)
 
     if (!COEFF_IS_MPZ(c1))      /* g is small */
     {
-        ulong u1 = (long) FLINT_ABS(c1);
-        ulong bits = (long) FLINT_BIT_COUNT(u1);
+        ulong u1 = FLINT_ABS(c1);
+        ulong bits = FLINT_BIT_COUNT(u1);
         if ((bits <= 1) || (exp * bits <= FLINT_BITS - 2))
         {
             fmpz_set_ui(f, n_pow(u1, exp));
